006data-collector/requirements.txt
Flask
Pillow
Flask-WTF
Flask-Bcrypt
Flask-Login
flask-loguru
Flask-SQLAlchemy
email-validator
python-dotenv
